Code cleanup & Filters FE
