+++ web/src/components/FOI/Home/ContextMenu.tsx
@@ -1,0 +1,208 @@
+import React, { useState } from 'react'
+import Popover from "@material-ui/core/Popover";
+import MenuList from "@material-ui/core/MenuList";
+import MenuItem from "@material-ui/core/MenuItem";
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faCirclePlus, faAngleRight } from '@fortawesome/free-solid-svg-icons';
+import { IconProp } from '@fortawesome/fontawesome-svg-core';
+import { savePageFlag } from '../../../apiManager/services/docReviewerService';
+import ConsultModal from "./ConsultModal";
+
+const ContextMenu = ({
+    requestId,
+    pageFlagList,
+    assignIcon,
+    openContextPopup,
+    setOpenContextPopup,
+    anchorPosition,
+    selectedPage,
+    selectedFile,
+    setPageFlagChanged
+}: any) => {
+
+    const [openModal, setOpenModal] = useState(false);
+    const [orgListAnchorPosition, setOrgListAnchorPosition] = useState<any>(undefined);
+    const [openConsultPopup, setOpenConsultPopup] = useState(false)
+    const [flagId, setFlagId] = React.useState(0);
+    const [docId, setDocumentId] = React.useState(0);
+    const [docVersion, setDocumentVersion] = React.useState(0);
+
+
+    const popoverEnter = (e: any) => {
+        setOrgListAnchorPosition(
+            e.currentTarget.getBoundingClientRect()
+        );
+        setOpenConsultPopup(true)
+    };
+
+
+    const addOtherPublicBody = (flagId: number, documentId: number, documentVersion: number) => {
+        setOpenModal(true);
+        setFlagId(flagId);
+        setDocumentId(documentId);
+        setDocumentVersion(documentVersion);
+    }
+
+    const closePopup = () => {
+        setOpenConsultPopup(false);
+    }
+
+    const savePageFlags = (flagId: number, pageNo: number, documentid: number, documentversion: number, publicbodyaction?: string, other?: string, programareaid?: number) => {
+        setOpenConsultPopup(false);
+        setOpenContextPopup(false);
+        savePageFlag(
+            requestId,
+            documentid,
+            documentversion,
+            pageNo,
+            flagId,
+            (data: any) => setPageFlagChanged(true),
+            (error: any) => console.log(error),
+            publicbodyaction,
+            other,
+            programareaid
+        );
+    }
+
+    const ministryOrgCodes = (pageFlag: any, documentId: number, documentVersion: number) => pageFlag.programareas?.map((programarea: any, index: number) => {
+        return (
+            <div key={programarea?.programareaid} onClick={() => savePageFlags(pageFlag.pageflagid, selectedPage, documentId, documentVersion, "", "", programarea?.programareaid)}>
+                <MenuList>
+                    <MenuItem>
+                        {programarea?.iaocode}
+                    </MenuItem>
+                </MenuList>
+            </div>
+        )
+    })
+
+    const otherMinistryOrgCodes = (pageFlag: any, documentId: number, documentVersion: number) => pageFlag.others?.map((other: any, index: number) => {
+        return (
+            <div onClick={() => savePageFlags(pageFlag.pageflagid, selectedPage, documentId, documentVersion, "", other)} key={index}>
+                <MenuList>
+                    <MenuItem>
+                        {other}
+                    </MenuItem>
+                </MenuList>
+            </div>
+        )
+    })
+
+    const showPageFlagList = () => pageFlagList?.map((pageFlag: any, index: number) => {
+        return (pageFlag?.name === 'Page Left Off' ?
+            <div className='pageLeftOff' onClick={() => savePageFlags(pageFlag.pageflagid, selectedPage, selectedFile.documentid, selectedFile.version)}>
+                <hr className='hrStyle' />
+                <div>
+                    {pageFlag?.name}
+                    <span className='pageLeftOffIcon'>
+                        <FontAwesomeIcon icon={assignIcon("Page Left Off") as IconProp} size='1x' />
+                    </span>
+                </div>
+            </div> :
+            <>
+                <MenuList key={pageFlag?.pageflagid}>
+                    <MenuItem>
+                        {(pageFlag?.name == 'Consult' ?
+                            <>
+                                <div onClick={popoverEnter}>
+                                    <FontAwesomeIcon style={{ marginRight: '10px' }} icon={assignIcon(pageFlag?.name) as IconProp} size='1x' />
+                                    {pageFlag?.name}
+                                    <span style={{ float: 'right', marginLeft: '51px' }}>
+                                        <FontAwesomeIcon icon={faAngleRight as IconProp} size='1x' />
+                                    </span>
+                                </div>
+                                <Popover
+                                    className='ministryCodePopUp'
+                                    id="mouse-over-popover"
+                                    anchorReference="anchorPosition"
+                                    anchorPosition={
+                                        orgListAnchorPosition && {
+                                            top: orgListAnchorPosition?.bottom,
+                                            left: orgListAnchorPosition?.right + 95,
+                                        }
+                                    }
+                                    open={openConsultPopup}
+                                    anchorOrigin={{
+                                        vertical: "center",
+                                        horizontal: "center",
+                                    }}
+                                    transformOrigin={{
+                                        vertical: "center",
+                                        horizontal: "center",
+                                    }}
+                                    onClose={() => closePopup()}
+                                    disableRestoreFocus
+                                >
+                                    <div className='ministryCodeModal' >
+                                        {ministryOrgCodes(pageFlag, selectedFile.documentid, selectedFile.version)}
+                                        {otherMinistryOrgCodes(pageFlag, selectedFile.documentid, selectedFile.version)}
+                                        <div className="otherOption" onClick={() => addOtherPublicBody(pageFlag.pageflagid, selectedFile.documentid, selectedFile.version)}>
+                                            <span style={{ marginRight: '10px' }}>
+                                                <FontAwesomeIcon icon={faCirclePlus as IconProp} size='1x' />
+                                            </span>
+                                            Add Other
+                                        </div>
+                                    </div>
+                                </Popover>
+                            </>
+                            :
+                            <div onClick={() => savePageFlags(pageFlag.pageflagid, selectedPage, selectedFile.documentid, selectedFile.version)}>
+                                <FontAwesomeIcon style={{ marginRight: '10px' }} icon={assignIcon(pageFlag?.name) as IconProp} size='1x' />
+                                {pageFlag?.name}
+                            </div>
+                        )}
+                    </MenuItem>
+                </MenuList>
+            </>
+        )
+    });
+
+    return (
+        <>
+            <Popover
+                anchorReference="anchorPosition"
+                anchorPosition={
+                    anchorPosition && {
+                        top: anchorPosition?.bottom,
+                        left: anchorPosition?.right,
+                    }
+                }
+                open={openContextPopup}
+                anchorOrigin={{
+                    vertical: "center",
+                    horizontal: "center",
+                }}
+                transformOrigin={{
+                    vertical: "center",
+                    horizontal: "center",
+                }}
+                onClose={() => setOpenContextPopup(false)}>
+                <div className='pageFlagModal'>
+                    <div className='heading'>
+                        <div>
+                            Export
+                        </div>
+                        <hr className='hrStyle' />
+                        <div>
+                            Page Flags
+                        </div>
+                    </div>
+                    {showPageFlagList()}
+                </div>
+            </Popover>
+
+            {openModal &&
+                <ConsultModal
+                    flagId={flagId}
+                    selectedPage={selectedPage}
+                    documentId={docId}
+                    documentVersion={docVersion}
+                    openModal={openModal}
+                    setOpenModal={setOpenModal}
+                    savePageFlags={savePageFlags} />
+            }
+        </>
+    );
+};
+
+export default ContextMenu;
