+++ web/src/components/FOI/Home/PageLeftOffModal.tsx
@@ -1,4 +1,3 @@
-import React, { useEffect, useState, useRef } from 'react'
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
@@ -12,12 +11,6 @@ const PageLeftOffModal = ({
     setOpenPageLeftOffModal,
     signout
 }: any) => {
-
-    // const signout = () => {
-    //     setOpenPageLeftOffModal(false);
-    //     localStorage.removeItem('authToken');
-    //     UserService.userLogout();
-    // }
 
     return (
         <div>
